chore: separate http concerns from sculpt_conn.c into sculpt_http.c
