+++ src/sculpt.h
@@ -42,6 +42,8 @@
 #define SC_BUFFER_OVERFLOW_ERR -17
 #define SC_MALFORMED_HEADER_ERR -18
 #define SC_CONN_CLOSE -19
+#define SC_HEADER_PARSE_ERR -20
+#define SC_HEADER_PARSE_INCOMPLETE_ERR -21
 
 #define SC_DEFAULT_BACKLOG 128
 #define SC_DEFAULT_EPOLL_MAXEVENTS 12
